# Provide the --fetch-age option for the RDS metrics collector