--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+- metrics-rds.rb: adding option `--fetch-age` to allow getting metrics in the past (@multani)
 
 ## [11.5.0] - 2018-05-17
 ### Added