@majormoses 2018-06-22
released: https://rubygems.org/gems/sensu-plugins-aws/versions/11.6.0